+++ src/kernel_evm/CHANGES.md
@@ -8,6 +8,8 @@
 
 ### Bug fixes
 
+- Increments the nonce of the transaction even if the transaction fails. (!9534)
+
 ### Breaking changes
 
 ### Internal